Fix warnings from curl 8.14
src/request.c:1158:19: error: call to ‘_curl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Werror=attribute-warning]

+++ inc/request.h
@@ -110,7 +110,7 @@ typedef struct RequestParams
     void *callbackData;
 
     // Request timeout. If 0, no timeout will be enforced
-    int timeoutMs;
+    long timeoutMs;
 } RequestParams;
